save work that does not work yet

+++ src/LikelihoodEvaluator.cpp
@@ -778,7 +778,7 @@ void LikelihoodEvaluator::fullOptimizeTreeinfo(pllmod_treeinfo_t *treeinfo)
   std::cout << "Genes in the tree: " << treeinfo->tip_count << std::endl;
   unsigned int it = 0;
   unsigned int max_it = 50;
-  unsigned int tolerance = 0.5;
+  double  tolerance = 0.5;
   do {
     previousLogl = newLogl;
     fullOptimizeTreeinfoIter(treeinfo);
